Rework ipify_test.go
The previous method of testing worked locally but is very dependent on the environment in which it runs. When running in a GitHub Action workflow, it apparently receives an IPv4 address when using the IPv6 endpoint, causing the test to fail.

The new method of testing removes any dependency on an external system and should therefore provide the same result everywhere.