build: add version number to release assets
The archive name template added by `goreleaser init` is great and all, but it does not include the release version in the file names, making it hard to see what version of the binary it contains.

The default name template used by goreleaser does include the version and is actually perfect.